1. 전 커밋에서 long double로 바꿨던 변수들 double로 다시 되돌림 2. Indicator.hpp에 Engines/BarData.hpp를 직접 include 하여 지표 개발 편의성 증대 3. 기타 코드 정리

+++ Includes/Indicators/SimpleMovingAverage.hpp
@@ -26,5 +26,5 @@ class SimpleMovingAverage final : public Indicator {
   size_t buffer_idx_;
 
   void Initialize() override;
-  Numeric<long double> Calculate() override;
+  Numeric<double> Calculate() override;
 };
