+++ Sources/Cpp/Indicators/Volume.cpp
@@ -1,11 +1,6 @@
 // 파일 헤더
 #include "Indicators/Volume.hpp"
 
-// 내부 헤더
-#include "Engines/BarData.hpp"
-#include "Engines/BarHandler.hpp"
-#include "Engines/BaseBarHandler.hpp"
-
 Volume::Volume(const string& name, const string& timeframe, const Plot& plot)
     : Indicator(name, timeframe, plot), symbol_idx_(-1) {}
 
@@ -14,6 +9,6 @@ void Volume::Initialize() {
   symbol_idx_ = bar_->GetCurrentSymbolIndex();
 }
 
-Numeric<long double> Volume::Calculate() {
+Numeric<double> Volume::Calculate() {
   return reference_bar_->GetBar(symbol_idx_, bar_->GetCurrentBarIndex()).volume;
 }
